refactor(app): rename autocomplete hooks to generic naming
